move cor functionality to common
