+++ profile-be-common/src/main/kotlin/com/lpfun/backend/common/cor/CorDslMarker.kt
@@ -1,0 +1,4 @@
+package com.lpfun.backend.common.cor
+
+@DslMarker
+annotation class CorDslMarker
