+++ profile-be-common/src/main/kotlin/com/lpfun/backend/common/cor/CorProcessor.kt
@@ -1,4 +1,4 @@
-package com.lpfun.backend.profile.domain.cor
+package com.lpfun.backend.common.cor
 
 class CorProcessor<T> private constructor(
     private val matcher: CorMatcherType<T>,
