+++ profile-be-common/src/main/kotlin/com/lpfun/backend/common/cor/IExec.kt
@@ -1,4 +1,4 @@
-package com.lpfun.backend.profile.domain.cor
+package com.lpfun.backend.common.cor
 
 interface IExec<T> {
     suspend fun execute(ctx: T)
