--- profile-domain/src/main/kotlin/com/lpfun/backend/profile/domain/cor/CorDslMarker.kt
@@ -1,4 +1,0 @@
-package com.lpfun.backend.profile.domain.cor
-
-@DslMarker
-annotation class CorDslMarker
