+++ profile-domain/src/main/kotlin/com/lpfun/backend/profile/domain/education/ProfileEducationCreateUseCase.kt
@@ -1,10 +1,10 @@
 package com.lpfun.backend.profile.domain.education
 
+import com.lpfun.backend.common.cor.IExec
+import com.lpfun.backend.common.cor.cor
 import com.lpfun.backend.common.model.profile.base.ProfileContextStatus
 import com.lpfun.backend.common.model.profile.base.stub.ProfileStubCreate
 import com.lpfun.backend.common.model.profile.education.ProfileEducationContext
-import com.lpfun.backend.profile.domain.cor.IExec
-import com.lpfun.backend.profile.domain.cor.cor
 
 class ProfileEducationCreateUseCase : IExec<ProfileEducationContext> {
     override suspend fun execute(ctx: ProfileEducationContext) = chain.execute(ctx.apply {
