+++ profile-domain/src/main/kotlin/com/lpfun/backend/profile/domain/skills/ProfileSkillsDeleteUseCase.kt
@@ -1,11 +1,11 @@
 package com.lpfun.backend.profile.domain.skills
 
+import com.lpfun.backend.common.cor.IExec
+import com.lpfun.backend.common.cor.cor
 import com.lpfun.backend.common.model.profile.base.ProfileContextStatus
 import com.lpfun.backend.common.model.profile.base.stub.ProfileStubDelete
 import com.lpfun.backend.common.model.profile.skills.ProfileSkillsAndTech
 import com.lpfun.backend.common.model.profile.skills.ProfileSkillsContext
-import com.lpfun.backend.profile.domain.cor.IExec
-import com.lpfun.backend.profile.domain.cor.cor
 
 class ProfileSkillsDeleteUseCase : IExec<ProfileSkillsContext> {
     override suspend fun execute(ctx: ProfileSkillsContext) = chain.execute(ctx.apply {
